+++ tests/end_to_end/misc1/irix-g-out.c
@@ -4,8 +4,8 @@ s32 test(s32 arg0, s32 arg1)
     s32 sp28;
     s32 sp24;
 
-    sp2C = (s32) ((D_4101C0 + (arg0 * 8))->unk4 + 1);
-    sp24 = (s32) (D_4101C0 + (arg0 * 8))->unk8;
+    sp2C = (D_4101C0 + (arg0 * 8))->unk4 + 1;
+    sp24 = (D_4101C0 + (arg0 * 8))->unk8;
     sp28 = func_00400174(1, 2, sp2C, arg1, arg0);
     if (sp28 == 0)
     {
